improve text legibility on ScreenGrooveStats login
Conditionally diffuse miso text to black when RainbowMode is in effect.

Change "Press &START; To Continue" text to use Wendy, similar to
ScreenEvaluationCasual.  Short of redesigning this screen for different
UX, I think this is the text we should be making large and bold, since
I've seen casual players get confused and walk away from the machine at
this screen.